requirements.txt
# CESD - Contrast-Enhanced Sparsified Decoding for VLM Hallucination Mitigation
# Python 3.10+ recommended, PyTorch 2.0+

# Core
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.25.0
safetensors>=0.4.0

# Vision-Language Models
Pillow>=10.0.0
qwen-vl-utils>=0.0.2  # For Qwen2-VL

# Evaluation
nltk>=3.8.0
spacy>=3.7.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Visualization & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
einops>=0.7.0
